Se mueve lógica de handleDowloadExcel a util
