+++ src/pages/Search/Search.js
@@ -12,11 +12,10 @@ import Filter from '../../components/Filter/Filter'
 // Utils
 import { handlePaste, handleKeyDown } from '../../utils/handleEvent'
 import { ORIGINS } from '../../utils/origins'
+import { handleDownloadExcel } from '../../utils/handleDowloadExcel'
 // Context
 import { AuthContext } from '../../context/authContext'
 import { BasicDataContext } from '../../context/basicDataContext'
-// XLSX
-import { utils, write } from 'xlsx'
 
 const Search = () => {
   const [reset, setReset] = useState(false) // Y esto?
@@ -58,28 +57,6 @@ const Search = () => {
       ...filters,
       [filterName]: value
     })
-  }
-
-  const handleDownloadExcel = (cases) => {
-    const workbook = utils.book_new()
-    const worksheet = utils.json_to_sheet(cases)
-
-    utils.book_append_sheet(workbook, worksheet, 'ResultCases')
-    const excelBuffer = write(workbook, {
-      bookType: 'xlsx',
-      type: 'array'
-    })
-
-    const data = new Blob([excelBuffer], {
-      type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-    })
-    const url = URL.createObjectURL(data)
-    const link = document.createElement('a')
-    link.href = url
-    link.setAttribute('download', 'listado-de-casos.xlsx')
-    document.body.appendChild(link)
-    link.click()
-    document.body.removeChild(link)
   }
 
   const handleReset = () => {
